Move navigation and header into the authenticated folder.
- These are not reusable components but specific to the id manager.
