+++ src/app/Authenticated/components/HeaderComponent.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
-import rifIdManager from '../../assets/images/rif-id-manager-gray.svg'
-import NetworkStatus from '../NetworkStatus/NetworkStatus'
-import ToolTip from '../Tooltip/Tooltip'
-import { truncateAddressDid } from '../../formatters'
+import rifIdManager from '../../../assets/images/rif-id-manager-gray.svg'
+import NetworkStatus from '../../../components/NetworkStatus/NetworkStatus'
+import ToolTip from '../../../components/Tooltip/Tooltip'
+import { truncateAddressDid } from '../../../formatters'
 
 interface HeaderComponentInterface {
   did: string | null
