+++ src/app/Dashboard/DashboardScreen.tsx
@@ -1,6 +1,4 @@
 import React from 'react'
-import HeaderComponent from '../../components/Header/HeaderComponent'
-import Navigation from '../../components/Navigation/Navigation'
 import IdentityInformationComponent from './panels/IdentityInformation'
 import { Authentication } from 'did-resolver'
 import Balance from './panels/Balance'
